[🧨refactor] : add globalStyle
