+++ src/App.tsx
@@ -1,6 +1,6 @@
 import AllContents from './components/AllContents/index';
 import Title from './components/Title';
-import { Container } from './style';
+import { Container, SpanStyle } from './style/style';
 function App() {
   const lastLogin = new Date().toString().slice(0, -21);
 
@@ -8,9 +8,7 @@ function App() {
     <>
       <Container>
         <Title />
-        <span style={{ color: 'white', marginBottom: '5px' }}>
-          Last login: {lastLogin}
-        </span>
+        <SpanStyle>Last login: {lastLogin}</SpanStyle>
         <AllContents />
       </Container>
     </>
